+++ Changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.2.1 (2023-11-24)
 
 * Fix: Prevent output incomplete expression like `1 : 2`
 * Doc: Clarify the operator precedence